codecs: fix pts truncation error
It was wrong to assume sizeof(pts) to be 4,
should use 8, instead.

needs corresponding host side change to fix
the bug completely.

BUG: 168593698

+++ system/codecs/omx/avcdec/MediaH264Decoder.cpp
@@ -158,7 +158,7 @@ h264_image_t MediaH264Decoder::getImage() {
         res.data = dst;
         res.width = *(uint32_t*)(retptr + 8);
         res.height = *(uint32_t*)(retptr + 16);
-        res.pts = *(uint32_t*)(retptr + 24);
+        res.pts = *(uint64_t*)(retptr + 24);
         res.color_primaries = *(uint32_t*)(retptr + 32);
         res.color_range = *(uint32_t*)(retptr + 40);
         res.color_trc = *(uint32_t*)(retptr + 48);
@@ -196,7 +196,7 @@ h264_image_t MediaH264Decoder::renderOnHostAndReturnImageMetadata(int hostColorB
         res.data = dst; // note: the data could be junk
         res.width = *(uint32_t*)(retptr + 8);
         res.height = *(uint32_t*)(retptr + 16);
-        res.pts = *(uint32_t*)(retptr + 24);
+        res.pts = *(uint64_t*)(retptr + 24);
         res.color_primaries = *(uint32_t*)(retptr + 32);
         res.color_range = *(uint32_t*)(retptr + 40);
         res.color_trc = *(uint32_t*)(retptr + 48);
